fix the search button design
