+++ src/index.tsx
@@ -1,4 +1,4 @@
-import React, { Suspense } from "react";
+import React from "react";
 import { createRoot } from "react-dom/client";
 
 import App from "./App";
@@ -6,17 +6,15 @@ import App from "./App";
 import "./index.css";
 import { Provider } from "react-redux";
 import store from "./redux/store";
-import Loader from "./components/Loader";
+
 
 const container = document.getElementById("root")!;
 const root = createRoot(container);
 
 root.render(
   <React.StrictMode>
-    <Suspense fallback={<Loader />}>
     <Provider store={store}>
       <App />
     </Provider>
-    </Suspense>
   </React.StrictMode>
 );
